requirements.txt
python-dotenv
pandas
numpy
setuptools
pymongo
certifi
pymongo[srv]
scikit-learn

## -e .